test: 변경된 패키지 구조에 맞게 테스트 구조 변경
